requirements.txt
beautifulsoup4>=4.12.2
Flask>=3.0.2
Flask-Migrate>=4.0.0
Flask-SQLAlchemy>=3.1.1
python-dateutil>=2.9.0.post0
requests>=2.31.0
spacy>=3.7.4
urllib3>=2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
APScheduler~=3.11.0
SQLAlchemy~=2.0.38
click~=8.1.8